Update lavfmuxer.cpp
This is to fix streams where the current code is not able to parse the stream and collect valid codec information as of invalid header. It tries to parse the stream until a valid packet is found and sample_rate is filled. Because if this is missing, the export crashes.
Might be improved.